Users can only use coupons 1 time, user select for coupons

+++ app/controllers/coupons_controller.rb
@@ -1,25 +1,26 @@
 class CouponsController < ApplicationController
 	def index
-		
+		redirect_to root_path, :alert => "Unauthorized" and return unless current_user.manager?
 		@coupons = Coupon.all.page(params[:page]).per(25).order('id ASC')
 	end
 
 	def new
+		redirect_to root_path, :alert => "Unauthorized" and return   unless current_user.manager?
 		@coupon = Coupon.new
 	end
 
 	def create
+		redirect_to root_path, :alert => "Unauthorized" and return   unless current_user.manager?
 		params[:coupon][:price] = params[:coupon][:price].to_s.gsub(',', '.').to_f unless params[:coupon][:price].blank?
-		params[:coupon][:category_ids] = params[:coupon][:category_ids].squish.split(" ").map(&:to_i)
-		params[:coupon][:part_ids] = params[:coupon][:part_ids].squish.split(" ").map(&:to_i)
-		params[:coupon][:user_ids] = params[:coupon][:user_ids].squish.split(" ").map(&:to_i)
+		params[:coupon][:category_ids] = params[:coupon][:category_ids].squish.split(" ").map(&:to_i).uniq
+		params[:coupon][:part_ids] = params[:coupon][:part_ids].squish.split(" ").map(&:to_i).uniq
+		params[:coupon][:user_ids] = params[:coupon][:user_ids].squish.split(" ").map(&:to_i).uniq
 		amount_days = params[:datetime][:days].to_i
     amount_hours = params[:datetime][:hours].to_i
     params[:coupon][:expiration_date] = Time.now.utc.to_date + amount_days.days + amount_hours.hours
 
 		@coupon = Coupon.new(coupon_params)
-
-		if @coupon.atleast_one_id(params[:coupon]) && @coupon.save
+		if @coupon.save
       redirect_to coupons_path
       flash[:success] = "Coupon added"
     else
@@ -28,6 +29,7 @@ class CouponsController < ApplicationController
 	end
 
 	def edit
+		redirect_to root_path, :alert => "Unauthorized" and return   unless current_user.manager?
 		@coupon = Coupon.find(params[:id])
 		weekday_array = ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"]
 		@day = @coupon.expiration_date.day.to_s
@@ -38,16 +40,17 @@ class CouponsController < ApplicationController
 	end
 
 	def update
+		redirect_to root_path, :alert => "Unauthorized" and return   unless current_user.manager?
 		params[:coupon][:price] = params[:coupon][:price].to_s.gsub(',', '.').to_f unless params[:coupon][:price].blank?
-		params[:coupon][:category_ids] = params[:coupon][:category_ids].squish.split(" ").map(&:to_i)
-		params[:coupon][:part_ids] = params[:coupon][:part_ids].squish.split(" ").map(&:to_i)
-		params[:coupon][:user_ids] = params[:coupon][:user_ids].squish.split(" ").map(&:to_i)
+		params[:coupon][:category_ids] = params[:coupon][:category_ids].squish.split(" ").map(&:to_i).uniq
+		params[:coupon][:part_ids] = params[:coupon][:part_ids].squish.split(" ").map(&:to_i).uniq
+		params[:coupon][:user_ids] = params[:coupon][:user_ids].squish.split(" ").map(&:to_i).uniq
 		amount_days = params[:datetime][:days].to_i
     amount_hours = params[:datetime][:hours].to_i
     params[:coupon][:expiration_date] = Time.now.utc.to_date + amount_days.days + amount_hours.hours
 
 		@coupon = Coupon.find(params[:id])
-    if @coupon.atleast_one_id(params[:coupon]) && @coupon.update(coupon_params)
+    if @coupon.update(coupon_params)
       redirect_to coupons_path
       flash[:success] = "Coupon updated"
     else
@@ -62,6 +65,7 @@ class CouponsController < ApplicationController
 	end
 
 	def destroy
+		redirect_to root_path, :alert => "Unauthorized" and return   unless current_user.manager?
 		coupon = Coupon.find(params[:id])
 		coupon.destroy
 		redirect_to coupons_path
